parser/htmlparser/tests/mochitest/test_xml_mislabeled_nobom.html

<!DOCTYPE html><meta charset=utf-8>
<title>Test for mislabeled XML entities</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css" />
<body>
<script class="testbody">
'use strict';

SimpleTest.waitForExplicitFinish();

var encodings = ['utf-8', 'uTf-8', 'UTF-8', 'utf-16', 'uTf-16', 'UTF-16'];
testxhr(encodings.shift());

function testxhr(encoding) {
  var xml = '<?xml version="1.0" encoding="' + encoding + '" ?>\n<test>test</test>';
  var xhr = new XMLHttpRequest();
  xhr.open('GET', 'data:text/xml,' + xml);
  xhr.send();
  xhr.onload = xhr.onerror = function(e) {
    ok(e.type === 'load', 'Test xhr loading for utf-8 XML labeled with ' + encoding);
    is(xhr.responseXML.documentElement.textContent, 'test',
       'Test xhr response for utf-8 XML labeled with ' + encoding);
    testiframe(encoding, xml);
  };
}

function testiframe(encoding, xml) {
  var iframe = document.createElement('iframe');
  iframe.src = 'data:text/xml,' + xml;
  iframe.onload = iframe.onerror = function(e) {
    ok(e.type === 'load', 'Test iframe loading for utf-8 XML labeled with ' + encoding);
    is(iframe.contentDocument.documentElement.textContent, 'test',
       'Test iframe content for utf-8 XML labeled with ' + encoding);
    if (encodings.length === 0) {
      SimpleTest.finish();
    } else {
      testxhr(encodings.shift());
    }
  };
  document.body.appendChild(iframe);
}
</script>
<div id="display"></div>
</body>
